Introduce git hook multiplexer
This change introduce git-hook-mux, a "multiplexer" for git hooks. The
concept is simple: the multiplexer is registered as a set of git hooks
(such as 'pre-commit'; see githooks(5)) and it will then take care to invoke a
user-defined set of hooks in turn. The user-defined hooks can be
configured using git-config. The section 'hook-mux' is searched for
values matching the commit hook the multiplexer got invoked for (i.e.,
the 'pre-commit' from before). All such configured hooks will be
executed, in order.
To configure a new hook it has to be configured, e.g.:
  $ git config --add "hook-mux.pre-commit" "<path-to-some-hook>"