Add config option to be more verbose
For debugging or informational purposes it might be relevant to see
some more information on that the hook multiplexer does and which hooks
it found.
For that matter, this change introduce a boolean config option,
hook-mux.verbose (defaulting to false; settable via git-config(1),
similar to the hook configuration) that can be used to increase
verbosity of the script.