Introduce -t/--hook-type parameter
There are cases when copying the script into each git hook directory is
not desirable and where an intermediate execution script disguises the
original file name. In such a case it is of importance to have other
means to convey the desired hook type to the script.
To that end, this patch introduces an optional command line parameter,
-t/--hook-type, that overrides any file name based hook type detection.