Introduce -s/--section parameter
With upcoming changes we will support nesting of the git-hook-mux. Such
an approach makes sense only if the configuration of each instance can
differ.
In order to enable such differing configuration, this change introduces
a new parameter, -s/--section, that can be used to specify the
git-config(1) section to read the configuration data (such as the hooks
to invoke) from.