Add <self> keyword to simplify nested invocation
We want to support nested usage scenarios of the script. That is, a
configured hook multiplexer should be allowed to contain another hook
multiplexer as one of its hooks.
Although such usage scenarios are possible, a problem is that we have no
way to refer to the currently executing git-hook-mux script. Hence, the
user is forced to specify the full absolute path to it -- which might be
impossible in case the script simply got copied into each repository.
This change introduces a keyword, '<self>', that, when encountered in a
hook command, is replaced with the current git-hook-mux instance
running.