Add explicit file management support
The goal of a of our hooks (especially commit hooks) is to inspect
files. Some of the hooks in turn might only want to look at a subset of
all files to be committed, e.g., only the Python files. Currently, each
and every hook is required to retrieve such a list of files itself which
defeats reusability and is error-prone because different approaches
for, say, determining the to-be-committed Python files, could result in
different file lists being created.
This change adds file support to the hook-mux script. Such support means
that the script accepts a list of files and passes this list on to the
hooks it invokes. Furthermore, the newly introduced --file-cmd parameter
can be used to specify a command that creates such a list of files or
filters a given list, respectively. This command is run before the hooks
are invoked. If the resulting file list is empty -- because no Python
files were found, for example -- the nesting ends and no further hook
multiplexers are invoked for this particular file list.